Add Track mesh construction capabilities.
Add lumps and pieces to be able to construct some blue prints to create
meshes in Blender: everything eventually will get optimized and phased out
since tl-tr-bl-br meshed can be constructed from a single g3d/g3dt mesh.

Add Zeni's test meshes, this is a fucking cool arc suitable as the starting
point "signal".

A Level object now manage creation and rendering of in-level static meshes
AND collision shapes.

MeshTypes are no longer a solution, MeshFactory exposes types but this
should be eliminated, in favour of Java's enums-sortof.

[fixed] UStillModel should pass the shader param to bind().

[add] Mipmap is now a configuration option, but not all assets are supposed
to reflect the setting (we love pixels!).