[fixed?] performance problem, code layout, refactoring.
The test_arch mesh causes the Asus Transformer to lose lot of
frames, but everything is fine if the EntityManager's SpriteBatch
is constructed with buffers, i suppose one single buffer is not
enough, but that's odd.
But excluding the cars and leaving the old SpriteBatch everything
then everything renders at full framerate (63/64fps).