Better shape for the light-in-shape bug.
The bug was caused by the car body's compound shape being composed by 4
attached shapes in a manner where the center of the car was being cut in
half by the polygon triangulator. Adjacent shapes were occluding the light.

The base/cars/electron.png asset need to be there in order to retain the
past defined shape (cannot copy it, nor move it, nor rename it in the
editor!)