Separate BeforeRender event from interpolation step
Separation is needed so that it can be performed before everythin else,
then the logic can be triggered to update before render, and finally the
rendering stage can perform the before render as well.
In this order, interpolables are ready, then logic can consume the data
and produce its output (such as camera zoom and position), finally the
renderer can perform the before render + render stage as well.

This architecture is still subject to change.