Add GhostCar event to signal replay finished, GameLogic changes
It wasn't GhostCar responsibility to remove its Replay or restart it, now
logic does it via the new raised event.
If no player is available, then the Replay will be restarted automatically
but this is subject to change.
Refactor GhostCar to accommodate changes, better code.
Fixed a CarHighliter problem where the alpha value was modulating the color
instead of the drawing opacity: it was working fine for the simple case but
conceptually it was the wrong implementation.
Minor refactoring.
Fixed DriftBar not being reset properly.
Fixed CarSimulator not resetting lateral forces properly (2 years old bug!)