Reworked quit logic
It works this way: a quit is signaled as soon as quit is requested (i.e.,
via 'Q' key) via a GameInput object; the current frame is still being
computed and rendered.
Quit pending is still being accounted for, the task manager will stop to
process tasks the next frame and as soon as the first tick can be signaled
(60 ticks per second, some tick/tickCompleted will NOT be called on faster
machines) the quit request is processed, a real game quit is scheduled and
a transition is setup.
This ends the game nicely, no more intra-frame updates.