Refactored alpha for GhostCar
Add tweenAlphaTo method to GhostCar to permit external logic to drive
and animate opacity directly.
Add method to know if the GhostCar instance is hinting at rendering
its depthmap for the SSAO pass or not.
Fixed set correct alpha opacity for ghosts/target when restarting all
replays.
GameWorldRenderer will not render SSAO for GhostCar instances not hinting
at it.